# swagger/swagger.yaml
swagger: '2.0'
info:
  version: "1.1.0"
  title: 'Open Library Project'
  description: 'Rest api for Library OpenClassroom Project'

# host: localhost:8080
# basePath: /Library
...
paths:
...
  /loans:
    get:
      tags:
      - "loan"
      summary: "Get a list of all loans"
      operationId: "getLoans"
      produces:
      - "application/json"
      parameters:
      - name: "status"
        in: "query"
        description: "Status values as filter in research"
        required: false
        type: "string"
        enum:
        - "loaned"
        - "returned"
        - "prolonged"
        - "late"
      - name: "user"
        in: "query"
        description: "User id as filter in research"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Loans found"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/LoanDto'
    post:
      tags:
      - "loan"
      summary: "Add a new loan"
      operationId: "addLoan"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "loanForm"
        description: "Loan that needs to be added to the database"
        required: true
        schema:
          type: "object"
          properties:
            copyId:
              type: "integer"
              format: "int64"
            userId :
              type: "integer"
              format: "int64"
      responses:
        201:
          description: "Loan added successfully to database"
          schema:
            $ref: '#/definitions/LoanDto'
        400:
          description: "Parameters are incorrect"
        403:
          description: "You are not allowed to perform this request"
  /loans/findLate:
    get:
      tags:
      - "loan"
      summary: "check and get loan list that are late"
      description: "trigger a checkup for all loan's and add late status to expired one's"
      operationId: "checkAndGetLateLoans"
      produces:
      - "application/json"
      parameters:
        - name: "Authorization"
          in: "header"
          description: "User identification"
          required: true
          type: "string"
      responses:
        200:
          description: "Loan found"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/LoanDto'
  /loans/{id}:
    get:
      tags:
      - "loan"
      summary: "Get loan by Id"
      operationId: "getLoan"
      produces:
      - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of loan to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Loan found"
          schema:
            $ref: '#/definitions/LoanDto'
        400:
          description: "Invalid ID supplied"
        404:
          description: "Loan not found"
    put:
      tags:
      - "loan"
      summary: "Update a loan by adding a status to it"
      operationId: "updateLoan"
      produces:
      - "application/json"
      consumes:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of loan to update"
          required: true
          type: "integer"
          format: "int64"
        - name: "status"
          in: "body"
          description: "Status values to add to loan history"
          required: true
          schema:
            type: "string"
            enum:
              - "loaned"
              - "returned"
              - "prolonged"
              - "late"
      responses:
        200:
          description: "Loan updated"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Loan not found"
  /loans/{id}/history:
    get:
      tags:
      - "loan"
      summary: "Get all loan's history"
      operationId: "getLoanHistory"
      produces:
      - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of loan"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Loan found"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/LoanStatusDto'
        400:
          description: "Invalid ID supplied"
        404:
          description: "Loan not found"
  /loans/{id}/extension:
    put:
      tags:
      - "loan"
      summary: "Extend an existing loan"
      operationId: "extendLoan"
      produces:
      - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of loan"
          required: true
          type: "integer"
          format: "int64"
        - name: "Authorization"
          in: "header"
          description: "User identification"
          required: true
          type: "string"
      responses:
        200:
          description: "Loan extended"
        400:
          description: "Invalid ID supplied"
        403:
          description: "Loan can't be extended"
        404:
          description: "Loan not found"
...
definitions:
  ReservationDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Reservation Id
      userId:
        type: integer
        format: int64
        description: User Id
      userEmail:
        type: string
        description: User email
      bookId:
        type: integer
        format: int64
        description: Book Id
      bookTitle:
        type: string
        description: Book title
      currentStatus:
        type: string
        description: Active Status for reservation
      currentStatusData:
        type: string
      statuses:
        type: array
        description: History of statuses
        items:
          $ref: '#/definitions/ReservationStatusDto'
  ReservationStatusDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Reservation Id
      status:
        type:  string
        description: status designation
      date:
        type: string
        description: date of the event
...
  BookDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Book Id
      isbn:
        type: string
        description: Book Isbn
      title:
        type: string
        description: Book title
      authors:
        type: array
        items:
          $ref: '#/definitions/AuthorDto'
      copiesAvailable:
        type: integer
        format: int64
        description: Number of copies available for loan
      reservationListFull:
        type: boolean
      reservations:
        type: array
        items:
          $ref: '#/definitions/ReservationDto'
  AuthorDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Author Id
      firstName:
        type: string
        description: Author first name
      lastName:
        type: string
        description: Author last name
      books:
        type: array
        items:
          type: string
          description: Book written by author
  BookCopyDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: BookCopy Id
      barcode:
        type: string
        description: BookCopy barCode
      editor:
        type: string
        description: Editor
      available:
        type: boolean
      book:
        $ref: '#/definitions/BookDto'
  LoanDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Loan Id
      userId:
        type: integer
        format: int64
        description: User Id
      userEmail:
        type: string
        description: User email
      startDate:
        type: string
        description: date of the loaning recording
      endDate:
        type: string
        description: date of the loaning end
      currentStatus:
        type: string
        description: name of the current status
      currentStatusDate:
        type: string
        description: date of the current status
      bookCopy:
        $ref: "#/definitions/BookCopyDto"
  LoanStatusDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Loan Id
      status:
        type:  string
        description: status designation
      date:
        type: string
        description: date of the event
securityDefinitions:
  basicAuth:
    type: basic
        
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /al26/Library/1.0.0
schemes:
 - https
 - http